Set GUI thread as the main thread. Added background operation with argument -d
